backend/requirements.txt
fastapi
uvicorn==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv
celery==5.4.0
redis==5.0.8
playwright==1.48.0
beautifulsoup4==4.12.3
torch==2.4.1
transformers==4.45.1
# tensorflow==2.17.0
numpy
pandas
resend==2.3.0
python-jose[cryptography]
passlib[bcrypt]
requests 
lxml
datasets 
matplotlib 
seaborn 
tqdm